add chart, auth

+++ src/pages/api/finance.ts
@@ -2,6 +2,7 @@ import { API, APIRequest, APIResponse } from "@/api";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { financeModule } from "@/database";
 import i18n from "@/i18n";
+import { token } from "@/token";
 
 export default async function handler(
   req: NextApiRequest,
@@ -9,6 +10,11 @@ export default async function handler(
 ) {
   if (req.method == "GET") {
     i18n.changeLanguage(req.query.lang as string || "en");
+    if (await token.checkToken(req.query.token as string) == null) {
+      API.failure(res, i18n.t("unauthorized"));
+      return;
+    }
+
     await financeModule.findAll().then((records) => {
       API.success(res, i18n.t("financeGetSuccess"), { records: records });
     }).catch((error: Error) => {
@@ -19,6 +25,11 @@ export default async function handler(
   else if (req.method == "POST") {
     const request: APIRequest = req.body;
     i18n.changeLanguage(request.lang || "en");
+    if (await token.checkToken(req.query.token as string) == null) {
+      API.failure(res, i18n.t("unauthorized"));
+      return;
+    }
+
     await financeModule.create({
       money: request.params.money,
       detail: request.params.detail,
@@ -31,6 +42,11 @@ export default async function handler(
   }
   else if (req.method == "DELETE") {
     i18n.changeLanguage(req.query.lang as string || "en");
+    if (await token.checkToken(req.query.token as string) == null) {
+      API.failure(res, i18n.t("unauthorized"));
+      return;
+    }
+
     const request: any = req.query;
     await financeModule.destroy({
       where: {
@@ -46,6 +62,11 @@ export default async function handler(
   else if (req.method == "PUT") {
     const request: APIRequest = req.body;
     i18n.changeLanguage(request.lang || "en");
+    if (await token.checkToken(req.query.token as string) == null) {
+      API.failure(res, i18n.t("unauthorized"));
+      return;
+    }
+
     await financeModule.update({
       money: request.params.money,
       detail: request.params.detail,
